Defconfig: Enable LLVM's ShadowCallStack and map virtually
ShadowCallStack (SCS) is an LLVM instrumentation mode that protects
against return address overwrites (like stack buffer overflows) by
saving a function's return address to a separately allocated ShadowCallStack
in the function prolog of nonleaf functions and loading the return
address from the ShadowCallStack in the function epilog.
The return address is also stored on the regular stack for compatibility
with unwinders, but is otherwise unused. This ensures that attacks that
modify the return address on the regular stack have no effect on program
control flow.

Enabling virtually mapped SCS provides better stack exhaustion protection
, but at the cost of per-thread memory consumption.